+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -484,73 +484,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("HomeViewSectionMarketingCollections", () => {
-    it("returns correct data", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-marketing-collections") {
-              __typename
-              component {
-                title
-              }
-
-              ... on HomeViewSectionMarketingCollections {
-                marketingCollectionsConnection(first: 2) {
-                  edges {
-                    node {
-                      title
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const collections = {
-        body: [
-          {
-            title: "Trending Now",
-          },
-          {
-            title: "Top Auction Works",
-          },
-        ],
-      }
-
-      const context = {
-        marketingCollectionsLoader: jest.fn().mockResolvedValue(collections),
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionMarketingCollections",
-          "component": Object {
-            "title": "Collections",
-          },
-          "marketingCollectionsConnection": Object {
-            "edges": Array [
-              Object {
-                "node": Object {
-                  "title": "Trending Now",
-                },
-              },
-              Object {
-                "node": Object {
-                  "title": "Top Auction Works",
-                },
-              },
-            ],
-          },
-        }
-      `)
-    })
-  })
-
   describe("HomeViewSectionViewingRooms", () => {
     it("returns correct data", async () => {
       const query = gql`
